Move URL fetch to ABIURLFetcher
