+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/abi/helpers/ABIURLFetcher.kt
@@ -1,0 +1,32 @@
+// SPDX-FileCopyrightText: 2026 Davide De Rosa
+//
+// SPDX-License-Identifier: GPL-3.0
+
+package com.algoritmico.passepartout.abi.helpers
+
+import java.io.IOException
+import java.net.HttpURLConnection
+import java.net.URL
+
+fun interface ABIURLFetcher {
+    fun fetch(url: String, cached: Boolean, timeout: Double): ByteArray
+
+    companion object : ABIURLFetcher {
+        override fun fetch(url: String, cached: Boolean, timeout: Double): ByteArray {
+            val connection = URL(url).openConnection() as HttpURLConnection
+            return try {
+                connection.requestMethod = "GET"
+                connection.useCaches = cached
+                connection.connectTimeout = timeout.toInt() * 1000
+                connection.readTimeout = timeout.toInt() * 1000
+                val status = connection.responseCode
+                if (status !in 200..299) {
+                    throw IOException("HTTP $status")
+                }
+                connection.inputStream.use { it.readBytes() }
+            } finally {
+                connection.disconnect()
+            }
+        }
+    }
+}
